feat: move about serverside
close #65
